LibVirt machinery should create the KVM domain
Currently, LibVirt machinery can determine the snapshot and revert
the VM state to it, but it jumps straight to waiting for a state
change without actually issuing a create call to the domain being
used. This is likely due to it expecting a live-state snapshot to
be instantiated which automatically starts the instance.

Add a conditional call self.vms[label].create() after resolving VM
snapshot state and prior to entering the wait state. If the VM
power state is still off, try to create the KVM domain with basic
error handling. This should not case race conditions as preceeding
snapshot operations are serialized and complete prior to returning.